add Processing-style constructor and build number
to support the Processing style paradigm of creating objects; although
the static initializeProcessing() may still be used, and in fact that
is what these new constructors do.

Also added xml file to properties of the eclipse project to generate a
build number automatically with each build.

(Sorry, code for both changes is interspersed in one commit.)